Refactor Patch Code select box handling
