+++ app/views/tenancies/index.html.erb
@@ -9,6 +9,24 @@
 
   <%= render 'tenancies/worktray/recommended_action_filter' if show_immediate_actions_filter? %>
 
+  <%= form_tag(worktray_path, :method => "get") do %>
+    <%= hidden_field_tag(:recommended_actions, params[:recommended_actions]) if params[:recommended_actions].present? %>
+    <%= hidden_field_tag(:paused, params[:paused]) if params[:paused].present? %>
+    <%= hidden_field_tag(:full_patch, params[:full_patch]) if params[:full_patch].present? %>
+    <table>
+      <tr>
+        <td>
+            <label class="govuk-label" for="code"><strong>Patch</strong><br/></label>
+            <%= select_tag(:patch, patch_codes_options(selected: params[:patch]), { :class => 'form-control' }) %>
+        </td>
+        <td>
+          <%= submit_tag('Filter by patch', class: 'button') %>
+        </td>
+      </tr>
+    </table>
+  <% end %>
+
+
   <% if @user_assigned_tenancies.none? %>
     <h3 class="tenancy_list tenancy_list__no_tenancies">No results found</h3>
