--- lib/hackney/income/patch_codes.rb
@@ -1,35 +1,0 @@
-module Hackney
-  module Income
-    class PatchCodes
-      def self.options
-        [
-          { name: 'All', code: nil },
-          { name: 'Arrears East Patch 1', code: 'E01' },
-          { name: 'Arrears West Patch 1', code: 'W01' },
-          { name: 'Arrears East Patch 2', code: 'E02' },
-          { name: 'Arrears West Patch 2', code: 'W02' },
-          { name: 'Arrears East Patch 3', code: 'E03' },
-          { name: 'Arrears West Patch 3', code: 'W03' },
-          { name: 'Arrears East Patch 4', code: 'E04' },
-          { name: 'Arrears West Patch 4', code: 'W04' },
-          { name: 'Arrears East Patch 5', code: 'E05' },
-          { name: 'Arrears West Patch 5', code: 'W05' },
-          { name: 'Arrears East Patch 6', code: 'E06' },
-          { name: 'Arrears West Patch 6', code: 'W06' },
-          { name: 'Arrears East Patch 7', code: 'E07' },
-          { name: 'Arrears West Patch 7', code: 'W07' },
-          { name: 'Arrears East Patch 8', code: 'E08' }
-        ]
-      end
-
-      def self.code_dropdown_options
-        res = []
-        options.each do |opt|
-          res << [opt[:name], opt[:code]]
-        end
-
-        res.sort_by { |i| i[0] }
-      end
-    end
-  end
-end
