+++ lib/hackney/income/use_case_factory.rb
@@ -105,10 +105,6 @@ module Hackney
         Hackney::Income::ActionDiaryEntryCodes
       end
 
-      def patch_codes
-        Hackney::Income::PatchCodes
-      end
-
       def find_or_create_user
         Hackney::Income::FindOrCreateUser.new(users_gateway: users_gateway)
       end
